Updated, looking at each gene at time
I have decided to concatenate all the gene matrices, filling up with
'NaNs' the places where the gene did not have a strain. Then I
calculated Genome Wide GRM (X*X'/M), decomposed this matrix and
calculated the pseudo-snps for each gene. Saved them in a folder and
then so far did a mantel test for genes that have the same size.