Fixed the normalization of the columns
After the skype meeting we had today. We have decided to, first of all
instead of creating nans matrix we created zero matrix and then we would
not need to correct for mean equal to zero (because snps were already
normalized). We normalized the data by column to have variance equal to
1, normalized the rows/individuals. Now the eigen decomposition of the
genetic matrix is happening.