Including gene locations based on pacbio data
This script will calculate the LD of all genes. In order to find the
gene lcoations I used SM158 as a reference. Blast this genome against
alignements of 10 strains of our data. Filtered the position of each
gene by the candidate that repeated most of the time (like if the
position was exact the same in the 10 blasts, 8 blasts and so on).
Sorted it first by plasmid origin and then by base pair.