pseudo kinship is fixed
The pseudo kinship is calculated based on the covariance matrix of the
pseudo snps. The calculation is based on the dot product of the pseudo
gene and dividing it by the number of strains in each gene. The final
Kinship matrix is then divided by half of the genes (not sure exactly
why).